rework darkmode and added dark-mpde.service, added dark mode for player
