+++ src/app/core/nav-bar/nav-bar.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { DarkModeService } from '../services/dark-mode-service/dark-mode.service';
 
 @Component({
   selector: 'app-nav-bar',
@@ -7,38 +8,11 @@ import { Component, OnInit } from '@angular/core';
 })
 export class NavBarComponent implements OnInit {
 
+  constructor(public darkmodeService: DarkModeService) {
+  }
+
   ngOnInit() {
-    const theme = localStorage.getItem('theme');
-    if (!theme) {
-      localStorage.setItem('theme', 'light');
-    } else {
-      this.setBodyTheme(theme);
-    }
-  }
-
-  changeThemeColor(){
-    const body = document.getElementsByTagName('body')[0];
-    if (body.classList.contains('dark-theme')){
-      body.classList.remove('dark-theme');
-      console.log('Dark removed');
-      localStorage.setItem('theme', 'light');
-
-    } else {
-      body.classList.add('dark-theme');
-      console.log('Dark added');
-      localStorage.setItem('theme', 'dark');
-    }
-    return true
-  }
-
-  private setBodyTheme(theme: string) {
-    const body = document.getElementsByTagName('body')[0];
-    body.classList.remove('light-theme', 'dark-theme');
-    body.classList.add(`${theme}-theme`);
-  }
-
-  isLightTheme() {
-    return localStorage.getItem('theme') === 'light';
+    this.darkmodeService.setCurrentTheme();
   }
 
 }
