+++ src/app/shared/audio-player/audio-player.component.html
@@ -13,7 +13,7 @@
   </div>
   <div class="row">
     <div class="col-md-3 col-12 order-md-1 order-2 mt-2">
-      <h3>PlayList</h3>
+      <h3 class="audio-player-h3">PlayList</h3>
       <ul class="list-group list-group-border">
         <li class="list-group-item list-group-item-border" [class.active]="i === activeItemIndex" [class.played]="isSongPlayed(item.id)"
           *ngFor="let item of audioList; let i = index" (click)="setActiveItem(i); openFile(item, i)">{{item.name}}</li>
@@ -22,7 +22,7 @@
     <div class="col-md-9 col-12 order-md-2 order-1">
       <div class="row">
         <div class="col-sm-8 col-md-5 col-lg-8 col-12 order-sm-1 order-2">
-          <h4> Book NameBook NameBook NameBook Name</h4>
+          <h4 class="audio-player-h4"> Book NameBook NameBook NameBook Name</h4>
         </div>
         <div class="col-sm-4 col-md-7 col-lg-4 col-12 order-sm-2 order-1">
           <div class="btn-group" dropdown>
