+++ src/styles.scss
@@ -17,9 +17,11 @@
 
   --base-dark-text-color: #fff;
   --base-dark-border-color: #373737;
-  --base-dark-main-color: #574ccf;;
+  --base-dark-main-color: #574ccf;
   --base-dark-secondary-color: #b2c0ff;
+  --base-dark-seek-color:#546ace;
 
+  --dark-mode-bg: #1e1e1e;
 }
 
 body {
